Added BuildRequests namespace and BuildRequestFulfiller
