+++ .gitignore
@@ -3,7 +3,7 @@
 */obj
 *.log
 
-Bot.sln.DotSettings.user
+Sajuuk.sln.DotSettings.user
 /.vs/Bot/v15
 /.vs
