+++ Sajuuk.Tests/Builds/BuildRequestFulfillerTests.cs
@@ -1,34 +1,35 @@
 ﻿using Moq;
 using Sajuuk.Builds;
+using Sajuuk.Builds.BuildRequests;
 using Sajuuk.GameData;
 using Sajuuk.GameSense;
 using Sajuuk.MapAnalysis;
 using Sajuuk.Tests.Fixtures;
 
-namespace Sajuuk.Tests;
+namespace Sajuuk.Tests.Builds;
 
-public class ControllerTests : IClassFixture<NoLoggerFixture> {
-    private readonly Mock<IUnitsTracker> _unitsTrackerMock = new Mock<IUnitsTracker>();
-    private readonly Mock<IBuildingTracker> _buildingTrackerMock = new Mock<IBuildingTracker>();
-    private readonly Mock<ITerrainTracker> _terrainTrackerMock = new Mock<ITerrainTracker>();
-    private readonly Mock<IRegionsTracker> _regionsTrackerMock = new Mock<IRegionsTracker>();
+public class BuildRequestFulfillerTests : IClassFixture<NoLoggerFixture> {
     private readonly TechTree _techTree = new TechTree(Mock.Of<IPrerequisiteFactory>()); // TODO GD TechTree should be mockable
     private readonly KnowledgeBase _knowledgeBase = new TestKnowledgeBase();
+    private readonly Mock<IUnitsTracker> _unitsTrackerMock = new Mock<IUnitsTracker>();
+    private readonly Mock<IBuildingTracker> _buildingTrackerMock = new Mock<IBuildingTracker>();
     private readonly Mock<IPathfinder> _pathfinderMock = new Mock<IPathfinder>();
-    private readonly Mock<IChatService> _chatServiceMock = new Mock<IChatService>();
-    private readonly Controller _controller;
+    private readonly Mock<ITerrainTracker> _terrainTrackerMock = new Mock<ITerrainTracker>();
+    private readonly Mock<IController> _controllerMock = new Mock<IController>();
+    private readonly Mock<IRegionsTracker> _regionsTrackerMock = new Mock<IRegionsTracker>();
 
-    public ControllerTests() {
-        _controller = new Controller(
-            _unitsTrackerMock.Object,
-            _buildingTrackerMock.Object,
-            _terrainTrackerMock.Object,
-            _regionsTrackerMock.Object,
+    private readonly BuildRequestFulfiller _buildRequestFulfiller;
+
+    public BuildRequestFulfillerTests() {
+        _buildRequestFulfiller = new BuildRequestFulfiller(
             _techTree,
             _knowledgeBase,
+            _unitsTrackerMock.Object,
+            _buildingTrackerMock.Object,
             _pathfinderMock.Object,
-            _chatServiceMock.Object,
-            new List<INeedUpdating>()
+            _terrainTrackerMock.Object,
+            _controllerMock.Object,
+            _regionsTrackerMock.Object
         );
 
         _unitsTrackerMock
@@ -41,7 +42,7 @@ public class ControllerTests : IClassFixture<NoLoggerFixture> {
     }
 
     [Fact]
-    public void GivenNoDronesAndBuildTypeBuild_WhenExecuteBuildStep_ThenNoProducersAvailable() {
+    public void GivenNoDronesAndBuildTypeBuild_WhenFulfillBuildRequest_ThenNoProducersAvailable() {
         // Arrange
         _unitsTrackerMock
             .Setup(unitsTracker => unitsTracker.OwnedUnits)
@@ -58,7 +59,7 @@ public class ControllerTests : IClassFixture<NoLoggerFixture> {
         );
 
         // Act
-        var result = _controller.FulfillBuildRequest(buildRequest);
+        var result = _buildRequestFulfiller.FulfillBuildRequest(buildRequest);
 
         // Assert
         var actualNoProducersAvailableFlag = result & BuildRequestResult.NoProducersAvailable;
@@ -66,7 +67,7 @@ public class ControllerTests : IClassFixture<NoLoggerFixture> {
     }
 
     [Fact]
-    public void GivenOneDroneAndBuildTypeBuild_WhenExecuteBuildStep_ThenNoProducersAvailable() {
+    public void GivenOneDroneAndBuildTypeBuild_WhenFulfillBuildRequest_ThenWillNotUseTheLastDrone() {
         // Arrange
         _unitsTrackerMock
             .Setup(unitsTracker => unitsTracker.OwnedUnits)
@@ -83,7 +84,7 @@ public class ControllerTests : IClassFixture<NoLoggerFixture> {
         );
 
         // Act
-        var result = _controller.FulfillBuildRequest(buildRequest);
+        var result = _buildRequestFulfiller.FulfillBuildRequest(buildRequest);
 
         // Assert
         var actualNoProducersAvailableFlag = result & BuildRequestResult.NoProducersAvailable;
@@ -91,7 +92,7 @@ public class ControllerTests : IClassFixture<NoLoggerFixture> {
     }
 
     [Fact]
-    public void GivenMoreThanOneDroneAndBuildTypeBuild_WhenExecuteBuildStep_ThenOk() {
+    public void GivenMoreThanOneDroneAndBuildTypeBuild_WhenFulfillBuildRequest_ThenOk() {
         // Arrange
         _unitsTrackerMock
             .Setup(unitsTracker => unitsTracker.OwnedUnits)
@@ -112,7 +113,7 @@ public class ControllerTests : IClassFixture<NoLoggerFixture> {
         );
 
         // Act
-        var result = _controller.FulfillBuildRequest(buildRequest);
+        var result = _buildRequestFulfiller.FulfillBuildRequest(buildRequest);
 
         // Assert
         var actualNoProducersAvailable = result & BuildRequestResult.NoProducersAvailable;
