+++ Sajuuk/Builds/BuildRequests/IFulfillableBuildRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace Sajuuk.Builds;
+﻿namespace Sajuuk.Builds.BuildRequests;
 
 public interface IFulfillableBuildRequest {
     public BuildType BuildType { get; }
