+++ Sajuuk/Builds/IBuildRequestFulfiller.cs
@@ -1,0 +1,7 @@
+﻿using Sajuuk.Builds.BuildRequests;
+
+namespace Sajuuk.Builds;
+
+public interface IBuildRequestFulfiller {
+    BuildRequestResult FulfillBuildRequest(IFulfillableBuildRequest buildRequest);
+}
