+++ Sajuuk/Debugging/BotDebugger.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using System.Numerics;
 using Sajuuk.ExtensionMethods;
-using Sajuuk.Builds;
+using Sajuuk.Builds.BuildRequests;
 using Sajuuk.Debugging.GraphicalDebugging;
 using Sajuuk.GameData;
 using Sajuuk.GameSense;
