+++ Sajuuk/Managers/EconomyManagement/TownHallSupervision/TownHallSupervisor.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using Sajuuk.ExtensionMethods;
-using Sajuuk.Builds;
+using Sajuuk.Builds.BuildRequests;
 using Sajuuk.Debugging.GraphicalDebugging;
 using Sajuuk.GameData;
 using Sajuuk.GameSense;
