+++ Sajuuk/Managers/ManagerFactory.cs
@@ -1,4 +1,5 @@
-﻿using Sajuuk.Builds;
+﻿using Sajuuk.Builds.BuildOrders;
+using Sajuuk.Builds.BuildRequests;
 using Sajuuk.Debugging.GraphicalDebugging;
 using Sajuuk.GameData;
 using Sajuuk.GameSense;
