+++ Sajuuk/Managers/ScoutManagement/ScoutingSupervision/ScoutSupervisor.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using System.Numerics;
 using Sajuuk.ExtensionMethods;
-using Sajuuk.Builds;
+using Sajuuk.Builds.BuildRequests;
 using Sajuuk.GameSense;
 using Sajuuk.Managers.ScoutManagement.ScoutingTasks;
