+++ Sajuuk/Managers/WarManagement/ArmySupervision/ArmySupervisor.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using System.Numerics;
 using Sajuuk.Algorithms;
-using Sajuuk.Builds;
+using Sajuuk.Builds.BuildRequests;
 using Sajuuk.GameSense;
 using Sajuuk.StateManagement;
