+++ Sajuuk/Managers/WarManagement/ArmySupervision/RegionalArmySupervision/RegionalArmySupervisor.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Sajuuk.ExtensionMethods;
 using Sajuuk.Algorithms;
-using Sajuuk.Builds;
+using Sajuuk.Builds.BuildRequests;
 using Sajuuk.Debugging.GraphicalDebugging;
 using Sajuuk.GameSense;
 using Sajuuk.Managers.WarManagement.ArmySupervision.UnitsControl;
