+++ Sajuuk/Managers/WarManagement/States/EarlyGame/EarlyGameBehaviour.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using Sajuuk.ExtensionMethods;
-using Sajuuk.Builds;
+using Sajuuk.Builds.BuildRequests;
 using Sajuuk.Debugging;
 using Sajuuk.Debugging.GraphicalDebugging;
 using Sajuuk.GameData;
