+++ Sajuuk/Managers/WarManagement/States/EarlyGame/EarlyGameBehaviourDebugger.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.Generic;
-using Sajuuk.Builds;
+using Sajuuk.Builds.BuildRequests;
 using Sajuuk.Debugging;
 using Sajuuk.Debugging.GraphicalDebugging;
 using Sajuuk.Managers.WarManagement.States.MidGame;
