+++ Sajuuk/Managers/WarManagement/WarManager.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.Generic;
-using Sajuuk.Builds;
+using Sajuuk.Builds.BuildRequests;
 using Sajuuk.Debugging.GraphicalDebugging;
 using Sajuuk.Managers.WarManagement.States;
 using Sajuuk.StateManagement;
